// apps/web/generated/api/models/ticketResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketPriority } from './ticketPriority';
import type { TicketStatus } from './ticketStatus';

export interface TicketResponse {
  ticket_id: string;
  ticket_code: string;
  created_by_user_id: string;
  target_user_id: string;
  subject: string;
  status: TicketStatus;
  priority: TicketPriority;
  closed_at: string | null;
  last_message_at: string | null;
}